+++ Model/RainBarrel.cs
@@ -2,18 +2,18 @@
 {
     public class RainBarrel : Container
     {
-        private const int CAPACITY_DEFAULT = 100;
-        private const int CAPACITY_SMALL = 80;
-        private const int CAPACITY_LARGE = 120;
+        private const double CAPACITY_DEFAULT = 100;
+        private const double CAPACITY_SMALL = 80;
+        private const double CAPACITY_LARGE = 120;
 
-        private RainBarrel(int capacity, int content) : base(capacity, content) { }
+        private RainBarrel(double capacity, double content) : base(capacity, content) { }
 
         public static RainBarrel Get()
         {
             return Get(0);
         }
 
-        public static RainBarrel Get(int content)
+        public static RainBarrel Get(double content)
         {
             return new RainBarrel(CAPACITY_DEFAULT, content);
         }
@@ -23,7 +23,7 @@
             return GetSmall(0);
         }
 
-        public static RainBarrel GetSmall(int content)
+        public static RainBarrel GetSmall(double content)
         {
             return new RainBarrel(CAPACITY_SMALL, content);
         }
@@ -33,7 +33,7 @@
             return GetLarge(0);
         }
 
-        public static RainBarrel GetLarge(int content)
+        public static RainBarrel GetLarge(double content)
         {
             return new RainBarrel(CAPACITY_LARGE, content);
         }
